move files to model package
