+++ api/src/main/java/bio/terra/tanagra/app/controller/EntitiesApiController.java
@@ -15,15 +15,15 @@ import bio.terra.tanagra.generated.model.ApiEnumHintValue;
 import bio.terra.tanagra.generated.model.ApiIntegerBoundsHint;
 import bio.terra.tanagra.generated.model.ApiListEntitiesResponse;
 import bio.terra.tanagra.generated.model.ApiRelationship;
+import bio.terra.tanagra.model.Attribute;
+import bio.terra.tanagra.model.DataType;
+import bio.terra.tanagra.model.Entity;
+import bio.terra.tanagra.model.Relationship;
 import bio.terra.tanagra.proto.underlay.EntitySearchHint;
 import bio.terra.tanagra.proto.underlay.EnumHint;
 import bio.terra.tanagra.proto.underlay.EnumHintValue;
 import bio.terra.tanagra.proto.underlay.FilterableAttribute;
 import bio.terra.tanagra.proto.underlay.IntegerBoundsHint;
-import bio.terra.tanagra.service.search.Attribute;
-import bio.terra.tanagra.service.search.DataType;
-import bio.terra.tanagra.service.search.Entity;
-import bio.terra.tanagra.service.search.Relationship;
 import bio.terra.tanagra.service.underlay.EntityFiltersSchema;
 import bio.terra.tanagra.service.underlay.Underlay;
 import bio.terra.tanagra.service.underlay.UnderlayService;
