+++ api/src/main/java/bio/terra/tanagra/service/query/api/ApiConversionService.java
@@ -3,10 +3,10 @@ package bio.terra.tanagra.service.query.api;
 import bio.terra.common.exception.NotFoundException;
 import bio.terra.tanagra.generated.model.ApiEntityDataset;
 import bio.terra.tanagra.generated.model.ApiEntityFilter;
+import bio.terra.tanagra.model.Attribute;
+import bio.terra.tanagra.model.Entity;
 import bio.terra.tanagra.service.query.EntityDataset;
 import bio.terra.tanagra.service.query.EntityFilter;
-import bio.terra.tanagra.service.search.Attribute;
-import bio.terra.tanagra.service.search.Entity;
 import bio.terra.tanagra.service.search.EntityVariable;
 import bio.terra.tanagra.service.search.Filter;
 import bio.terra.tanagra.service.underlay.Underlay;
