+++ api/src/main/java/bio/terra/tanagra/service/query/api/ExpressionConverter.java
@@ -3,9 +3,9 @@ package bio.terra.tanagra.service.query.api;
 import bio.terra.common.exception.BadRequestException;
 import bio.terra.tanagra.generated.model.ApiAttributeValue;
 import bio.terra.tanagra.generated.model.ApiAttributeVariable;
-import bio.terra.tanagra.service.search.Attribute;
+import bio.terra.tanagra.model.Attribute;
+import bio.terra.tanagra.model.DataType;
 import bio.terra.tanagra.service.search.AttributeVariable;
-import bio.terra.tanagra.service.search.DataType;
 import bio.terra.tanagra.service.search.EntityVariable;
 import bio.terra.tanagra.service.search.Expression;
 import bio.terra.tanagra.service.underlay.Underlay;
