+++ api/src/main/java/bio/terra/tanagra/service/search/EntityVariable.java
@@ -1,5 +1,6 @@
 package bio.terra.tanagra.service.search;
 
+import bio.terra.tanagra.model.Entity;
 import com.google.auto.value.AutoValue;
 
 /**
