+++ api/src/main/java/bio/terra/tanagra/service/search/Expression.java
@@ -1,5 +1,6 @@
 package bio.terra.tanagra.service.search;
 
+import bio.terra.tanagra.model.DataType;
 import com.google.auto.value.AutoValue;
 
 /** A construct in a query syntax tree that evaluates to a value. */
