+++ api/src/main/java/bio/terra/tanagra/service/underlay/Column.java
@@ -1,6 +1,6 @@
 package bio.terra.tanagra.service.underlay;
 
-import bio.terra.tanagra.service.search.DataType;
+import bio.terra.tanagra.model.DataType;
 import com.google.auto.value.AutoValue;
 
 /** A SQL column within a table in an underlay. */
