+++ api/src/main/java/bio/terra/tanagra/service/underlay/Underlay.java
@@ -1,8 +1,8 @@
 package bio.terra.tanagra.service.underlay;
 
-import bio.terra.tanagra.service.search.Attribute;
-import bio.terra.tanagra.service.search.Entity;
-import bio.terra.tanagra.service.search.Relationship;
+import bio.terra.tanagra.model.Attribute;
+import bio.terra.tanagra.model.Entity;
+import bio.terra.tanagra.model.Relationship;
 import com.google.auto.value.AutoValue;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
