+++ api/src/test/java/bio/terra/tanagra/service/query/api/ApiConversionServiceTest.java
@@ -16,10 +16,10 @@ import bio.terra.tanagra.generated.model.ApiBinaryFilterOperator;
 import bio.terra.tanagra.generated.model.ApiEntityDataset;
 import bio.terra.tanagra.generated.model.ApiEntityFilter;
 import bio.terra.tanagra.generated.model.ApiFilter;
+import bio.terra.tanagra.model.DataType;
 import bio.terra.tanagra.service.query.EntityDataset;
 import bio.terra.tanagra.service.query.EntityFilter;
 import bio.terra.tanagra.service.search.AttributeVariable;
-import bio.terra.tanagra.service.search.DataType;
 import bio.terra.tanagra.service.search.EntityVariable;
 import bio.terra.tanagra.service.search.Expression;
 import bio.terra.tanagra.service.search.Filter;
