+++ api/src/test/java/bio/terra/tanagra/service/query/api/ExpressionConverterTest.java
@@ -9,8 +9,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import bio.terra.common.exception.BadRequestException;
 import bio.terra.tanagra.generated.model.ApiAttributeValue;
 import bio.terra.tanagra.generated.model.ApiAttributeVariable;
+import bio.terra.tanagra.model.DataType;
 import bio.terra.tanagra.service.search.AttributeVariable;
-import bio.terra.tanagra.service.search.DataType;
 import bio.terra.tanagra.service.search.EntityVariable;
 import bio.terra.tanagra.service.search.Expression;
 import bio.terra.tanagra.service.search.Variable;
