+++ api/src/test/java/bio/terra/tanagra/service/search/DataTypeVisitorTest.java
@@ -2,6 +2,7 @@ package bio.terra.tanagra.service.search;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import bio.terra.tanagra.model.DataType;
 import bio.terra.tanagra.service.underlay.NauticalUnderlayUtils;
 import bio.terra.tanagra.service.underlay.Underlay;
 import org.junit.jupiter.api.Tag;
